requirements.txt
streamlit
langgraph
langchain
langchain-openai
langchain-groq
langchain-community
pyalex
faiss-cpu
pymupdf4llm
numpy
pandas
python-dotenv
pydantic
sentence-transformers
thefuzz
python-Levenshtein
